Added Header files for the Player class
Creating Header and Source Files for the Player Class
