+++ Player.hpp
@@ -1,0 +1,20 @@
+#include "PokemonType.hpp"
+#include "PokemonChoice.hpp"
+#include "Utility.hpp"
+#include <string>
+using namespace std;
+
+class Player 
+{
+    public:
+    string name;
+    Pokemon chosenPokemon;
+
+    // Default Constructor
+    Player();
+    // Parameterized Constructor
+    Player(string p_name, Pokemon p_chosenPokemon);
+    // Copy Constructor 
+    Player(const Player &other);
+    void ChoosePokemon(int choice);
+};
